+++ r2c_isg/structures/__init__.py
@@ -1,4 +1,3 @@
-from .dataset import Dataset
 from .projects import Project, DefaultProject, project_map
 from .versions import Version, DefaultVersion, version_map
