+++ r2c_isg/util.py
@@ -1,7 +1,7 @@
 import subprocess
 import traceback
 
-from r2c_isg.structures import Dataset
+from r2c_isg.dataset import Dataset
 
 
 def get_name():
